Support --add-branch-tree-prefix option
This option detects branching a subdirectory of an existing branch.
If not enabled, the subdirectory will become a root directory of the
new branch, and then git will have to properly detect renames.
If enabled, file on the new branch will be added a directory prefix,
to match their position in the parent