Serialize the log output
With the dependency scheme introduced for child branches,
the commits are made out of order with the revisions, which
makes analysing the log difficult.

This change writes the log in the order of revisions. Commits are
logged in the revision order.

This change also assists for making commits asynchronously,
and for invoking "git log" asynchronously after the commit was made.