Use commit dependencies to serialize execution
Make branches and commits dependent objects.
Make commits only when all dependent objects are ready.

This prepares for the child merged branches feature,
and for making commits asynchronously with other processing

Hold a branch revision 'not ready' until the next revision is processed on this branch

This is needed to properly process SkipCommit specification
The program needs to hold finalizing a previous
commit until the new revision is processed on the given branch.